feat: add safetycontroller

+++ src/InfoPlottor.py
@@ -37,7 +37,7 @@ class InfoPlottor:
         time.sleep(0.3)
         assert self.plot_info_thread.is_alive() is False
 
-    def set_target_pos(self, target_pos: list[int]) -> None:
+    def set_target_pos(self, target_pos: list[float]) -> None:
         self.target_pos = target_pos
 
     def _plot_info(self):
